Step Selector: Allow mouse to be anywhere while dragging. (#5809)
Motivation for features / changes
It was not a good experience to force users to keep mouse on axis while dragging the fob. This allows users to mouse mouse anywhere while the fob is being dragged and it will continue dragging until a mouse up event.

Technical description
The main changes for this are in the card_fob_controller_component.ts file. It is difficult to register for document level mouse events from the html page so we have to do it programatically. Also, the reference for this listener is held on the document object. This causes some issues as it is never removed even when the component is destroyed. Also since the call comes from that object it does not hold a reference to "this". To keep the reference I bound "this" to the call. I also remove the listener when dragging stops so we do not keep adding more and more listeners to the body. This turned out to be a bit tricky as we had to keep a reference to the EventListener and calls to .bind always create a new reference. I had to make that reference of type any to avoid issues with --strictFunctionTypes as described nicely in this stack overflow answer.

Also of note: mouse events registered to the document will continue firing even when the pointer is outside the browser window. Therefore there is no longer a need for mouseleave events as we will continue dragging even off the page and the mouseup will always fire no matter where the mouse is when the user releases their click.

googlers see b/239453022

This also solves the issue that the dragging does not work when steps are too far apart. Googlers see b/238666730