Step Selector: Disable selector and table when Horizontal Axis is not set to Step. (#5827)
Motivation for features / changes
We have not implemented all that is needed to make this feature work for other types of axis. Instead of implementing this in Phase 1 we are just disabling this when the axis is not set to Step.

Technical description of changes
I basically just utilized the getMetricsXAxisType selector to disable and enable the proper pieces. This started to make the calculation a bit complex to have directly in the html file so I often moved it to a utility function in the component. I also utilized to "title" attribute to give a tooltip which hopefully will alleviate any confusion users may have.

Screenshots of UI changes
Here is what the settings looks like when disabled and the mouse if hovering over the step selector setting.

Screen Shot 2022-07-28 at 5 31 20 PM

Alternate designs / implementations considered
I considered just implementing the pieces to enable relative time as well. The problem is implementing a way to allow typing the time.