Add tooltip to run selection to hint about double click. (#5842)
We recently added the ability to double click a run checkbox to select it only and deselect the other runs.

User feedback (googlers, see b/238365158#comment9) said it was not discoverable. 

This change adds a tooltip to the checkbox, which includes a hint about double click. I use "title" instead of "matTooltip" just because we have many icon buttons that also use this and the behavior is more consistent with those buttons.