Enable Time Series dashboard when only image events are available. (#5852)
The Time Series dashboard was not appearing for TensorBoards with only image events.

We suspect this was just an oversight when Images were added to the Time Series dashboard. There is a function in the Metrics (aka "Time Series") plugin, data_plugin_names() which lists all data types for which the plugin should be enabled. It currently lists scalars and histograms but does not list images. This change just adds images to that list.