Feature Helper: add action and effect for storing flags (#5765)
Motivation
The goal here is to add a page which helps develops set feature flags. Part of the features to this page is that When you set a flag in this page it will persist. This PR adds the functionality to store flag status in localStorage to the featureFlagDataSource. It also adds the action and effect used to store the flags.