linked time: Add value change to data table in range selection (#5851)
Motivation for features / changes
This is the first change which adds a column in the data table for range selection. We plan to have many columns which deal with a range for the launch of linked time.

Technical description of changes
This change sets the structure for changing the column headers based on the linked time selection and the structure for having different styles in certain columns of the data table. I used ngSwitch for the style in anticipation for more styles in other columns.