Stop event propagation when starting to drag step selector (#5858)
Motivation for features / changes
Some strange behavior was observed when trying to drag the step selector. The root cause of this behavior was that when clicking and dragging in certain places the mouse was triggering other events(such as trying to move an entire graphic or highlighting text. This PR stops the propagation of those mouse clicks so that only the dragging that we want to occur occurs.