Improve FeatureFlagMetadata type so parse only specified when required. (#5857)
In the newly-added FeatureFlagMetadataMap, we are specifying parseValue functions for a number of feature flags that don't need them.

This change improves the typing of FeatureFlagMetadata so we only specify parseValue functions when they are required - that is, when the metadata includes a queryParamOverride value.

This makes the type a bit more strict. This allows us to get rid of the parseBooleanOrNull parser.