Use FeatureFlagMetadata from state in DeepLinkProviders. (#5877)
Now that we can consume FeatureFlagMetadata from state. (See #5876 and, Googlers, cl/467745607), we can start using it.

We update DashboardDeeplinkProvider to use metadata from FeatureFlagState instead of via constructor. The constructor argument is left optional for the time being, until we can update internal code to remove usage.

We also make the metadata field mandatory now that it is specified universally.