Step Selector: Add layer to stop useless re-render (#5887)
Motivation for features / changes
Since the getTimeSelectionTableData() function returned a new object to the data-table the entire table re-rendered every time the mouse moved over the scalar chart. This was causing performance issues. In this change we added a component layer which essentially takes the only the inputs required for the data table. This way the re-rendering only happens when those inputs change.

Detailed steps to verify changes work correctly (as executed by you)
Using the flash painting tool in the rendering chrome dev tools I could prove that the table was no longer re-rendering when the mouse moves over the chart.

Alternate designs / implementations considered
Another option is to just return the same object from getTimeSelectionTableData() when nothing has changed. This also increases performance. However, it was decided using angular tools for this was the correct way to fix the performance issue.

googlers see: b/242572017