Limit feature flags sent to server. (#5882)
In all requests from the TensorBoard application to the HTTP server we currently populate the "X-TensorBoard-Feature-Flags" header with the entire set of feature flags and values. We decided we only want to send the set of feature flags that are relevant to the HTTP Server and whichever DataProvider implementation backs it.

This change limits the feature flags in two ways:

1. Introduce the "sendToServerWhenOverridden" metadata field. Only flags that are marked as "sendToServerWhenOverridden" can be sent to the HTTP Server.
2. Only send flags which have been overriden by the user.

So for a flag value to be included in "X-TensorBoard-Feature-Flags" it must both have "sendToServerWhenOverridden" set to true and it must be overridden by the user.

Note that none of the flags in Standalone TensorBoard are marked as sendToServerWhenOverridden - the use case is currently internal.