linked time: Add min, max, and percentage change columns (#5884)
Motivation for features / changes
When we launch linked time we will be allowing a range selection. The data table needs to be able to handle range values when in that mode. This change adds the Min, Max, and Percentage Change columns so the user knows which step is selected for each run.

Technical description of changes
There are two main changes between these columns and others that have been added. The first is that the min and max have some calculation that I included in a utility function that I put directly in the ScalarCardDataTable. The second is that the percentage change uses arrows similar to the value change column. For code reuse I created an ng-template for the arrow and used that template for both columns.