Fix brittle Zlib-dependent test
Tests that expect zlib to produce a particular bitstream for some input
data are brittle, because the output of zlib is variable for different
implementations and different tunings.

We trust the underlying zlib `encode_png` op does the right thing if it
has a valid PNG header and is of a reasonable size.