linked time: updates linked time selection to latest step selector update (#5891)
The values update between step selector and linked time have been an hidden and ambiguous part. They are largely undependable prior to this PR: (1)updates on step selector will not reflect on linked time (2)updates to linked time will only update to a single step selector only when users dragging on that fob.
In FE team we made a preliminary decision to change case (1): the latest update on step selector will be remembered and set as the linked time selection (as opposed to always set to min value in the dashboard). case (2) is still true.