Linked Time: Clean data table headers (#5899)
Motivation for features / changes
In our next phase of launching linked time we introduce the range selection. This has a whole new set of columns in the data table. These new columns have already been added however, after discussions with the team and with users we want to change the ordering and text in the headers. This PR makes those changes to the headers.

Technical description of changes
This change is mostly text changes and some css. The only real logic added was to add an ngswitch inside the data_table_component.ng.html file which adds the triangle mat-icon(which is called "change_history" for some reason). This is also the first time we added that icon so I had to add it to our build in the js.bzl file.