Linked Time: Clean up slider (#5909)
Motivation for features / changes
While messing around with the slider I found some issues. The propagation of the event caused some problems so I stoped that. Also, the change in https://github.com/tensorflow/tensorboard/pull/5908 makes it unclear if you are dragging the thumb when there are few steps. Changing the cursor to a grabber helps give that indication and also helps range slider align with the single slider.