Linked Time: Round to nearest integer when dragging range slider (#5908)
Motivation for features / changes
Steps are discrete. It does not make sense to have a slider which chooses steps to use floating point. This forces that slider to round to the nearest integer.

Technical description of changes
Since the slider is a widget I wanted to allow the user of the widget to decide if floating point was something that they wanted. Therefore I made it an optional input.