Allow event propogation on Fob clicks (#5907)
Motivation for features / changes
Event propagation was causing some when dragging the line above the fob. This prompted us to stop propagation for startDrag events. However, that propagation is needed on the fob to allow users to type in the step. This change adds logic to allow propagation of mouse events on fob clicks.