Linked Time: Sort Data Table based on header clicks (#5911)
Motivation for features / changes
With the introduction of the data table in Scalar Cards many more useful features become available. One of these features is sorting that data table. This change adds the ability for users to sort by any column by clicking on the header to that column.

Technical description of changes
The sortingInfo object, which holds the information about which column and what order we are sorting, is stored in the ScalarCardComponent. It is currently not put into the ngrx state. It is piped through the ScalarCardDataTable and into the DataTableComponent. The ScalarCardDataTable uses this information to sort the data passed to the DataTableComponent and the DataTableComponent itself uses the information to update the header UI and inform the information emitted when the header is clicked.