Linked Time: Use TimeSelectionWithAffordance for onLinkedTimeSelectionChanged event (#5916)
Motivation for features / changes
Recent changes broke a feature in Histogram Cards which allowed users to enable range selection by clicking on the chart. This fixes that breakage.

Technical description of changes
The onLinkedTimeSelectionChanged event was still emitting a simple TimeSelection object. However, the function being called on that event was expecting a TimeSelectionWithAffordance object. This changes the object to emit a TimeSelectionWithAffordance and created a new type of affordance to use in this event.