Linked Time: Round to next step in fob controller (#5918)
When zooming on the scalar card the min and max values become floating points. Then when you drag the fob passed these values it sets those fobs to floating point steps. Since steps are discrete this is unwanted behavior. This change ensure that we show the min or max step instead of the value(often floating point) at the edge of the chart.