Linked Time: Sorting Header css cleanup and hover indication (#5915)
Motivation for features / changes
This is a followup to PR Linked Time: Sort Data Table based on header clicks #5911 which added the sorting functionality to the data table. This is purely a UI change to help indicate to users that you can click the header to sort and help make the feature look better.