Linked Time: Add Range Checkbox behind new flag (#5919)
Motivation for features / changes
We plan to launch a new feature which allows users to enable a range instead of a single for scalar cards. This is the first PR towards that goal.

Technical description of changes
This PR adds a feature flag and then uses that flag to enable or disable a checkbox in the settings panel. The checkbox currently has no functionality. That will come in later PRs.