Linked Time: Disable Step Selection when dismissing single fob (#5917)
Motivation for features / changes
When the fobs are dismissed you expect the fob to disappear. However, when "Link by step" is enabled it only goes into "step selection" mode which still has the fob in Scalar Cards. This change make the dismissal disable step selection so that the fob always disappears. For consistency this also disables step selection when dismissing For on Histogram Card.