Bug Fix: Fix data table sorting of nans (#5953)
* Motivation for features / changes
Sorting in a column which contains `NaN` in the data table used to break
the sorting.

* Technical description of changes
Sorting was broken because `NaN` values were actually being passed in as
strings which lead to all the values in the column being sorted as
though they were strings. To resolve this I translate all values into
"sortable" versions which serves to additionally guard against other
potentially problematic types.

* Detailed steps to verify changes work correctly (as executed by you)
1) Start tensorboard with a logdir which contains runs with `NaN`
values.
2) Go to localhost:6006?enableDataTable
3) Enable step selection and select a step in a scalar card where some
but not all values are `NaN`
4) Sort the table by the `value` column and ensure that sorting makes
sense.

* Alternate designs / implementations considered
We could run all incoming series data through a map and convert the
string `"NaN"` to `NaN` so sorting would work. That solution seems a bit
hackier though and doesn't do as much to guard against other sorting
anomalies.