Step Selector: Stop modifying timeSelection Input (#5947)
* Motivation for features / changes
We used to modify the timeSelection Input object in the
ScalarCardComponent. Since #5941 This object is now updated in the
ScalarCardContainer as it should be. However, the modifications in lower
level components are still there. This change removes those
modifications and completely relies on the container changes.

* Technical description of changes
I also took the opportunity to do some cleanup that I thought was nice.
I moved the lineChartZoom$ declaration down to the other BehaviorSubject
definitions and defined it as readonly. I also made the minMaxStep
Observable a BehaviorSubject so that I can use the value to check the
new timeSelections before modifying it. This check was previously done
in the ScalarCardComponent since we did not have access to the
minMaxStep values in the container. The logic makes more sense in the
Container and allows us to remove that function entirely and just emit
the event directly.