Added section on logging batch-level metrics to get_started.ipynb (#5928)
* Motivation for features / changes
This PR adds a new section to "Get Started with Tensorboard" focused on
adding custom metrics for batch-level logging during training. This is
related to an issue (b/206800675) with the functionality of the
update_freq parameter that causes the example code in TF docs
(https://www.tensorflow.org/api_docs/python/tf/keras/callbacks/TensorBoard)
to not work as intended. Adding this section to the notebook would
provide a good workaround to allow custom batch-level summaries for both
subclassed and Functional API Models.

* Technical description of changes

* Screenshots of UI changes
![Screen Shot 2022-09-15 at 1 41 20 PM
(2)](https://user-images.githubusercontent.com/60985914/190504933-34a79065-2e53-4864-a86e-38a1bcddd8bd.png)
![Screen Shot 2022-09-15 at 1 45 00 PM
2](https://user-images.githubusercontent.com/60985914/190505459-670cc3ef-6a8c-41e1-a626-13977fa9e4c2.png)

* Detailed steps to verify changes work correctly (as executed by you)
Results in Tensorboard are as intended. A full run-through was executed
for the notebook to ensure existing and new functionality.

* Alternate designs / implementations considered
Alternates included moving the new section up to after Keras model.fit()
section, but I believe new section should be last, as it should be after
the introduction of using tf.summary in the "Other methods" section.

Co-authored-by: Riley Jones <78179109+rileyajones@users.noreply.github.com>