Update gif for tensorflow.org/tensorboard with Time Series as default dashboard. (#5860)
With the TensorBoard 2.10 release the "Time Series" dashboard will be
the default dashboard for logs with scalars, histograms, or images
events. We are updating external-facing documentation to reflect this
change.

This change updates the gif that shows up on
http://tensorflow.org/tensorboard. The previous gif highlighted the
Scalars, Images, Graphs, Histograms, and Distributions dashboards. The
new gif highlights the Time Series dashboard -- showing scalars then
images then histograms using the data toggle -- and the Graphs
dashboard. Notably, the Distributions dashboard is not present in the
new gif.

How I generated the gif:
* Use my Mac's native recording capabilities (cmd+shift+5) at 1200x900
to generate a mov file.
* Transfer the mov file to my linux machine and run ffmpeg:
* `ffmpeg -i tensorboard.mov -vf
"fps=7,scale=800:-1:flags=lanczos,split[s0][s1];[s0]palettegen[p];[s1][p]paletteuse"
-loop 0 tensorboard.gif`
  * Use 7 FPS to be similar to the previous gif.
* Scale down to 800x600 to be similar to the previous gif (it was
800x605).
* Use a custom pallette (palettegen and paletteuse) to reduce dithering.
See
https://superuser.com/questions/556029/how-do-i-convert-a-video-to-gif-using-ffmpeg-with-reasonable-quality.
  * Configure to have an infinite loop.

The resulting gif is about 8 seconds and 437KB in size.

Note that I originally tried to create a gif that scrolls down the Time
Series dashboard but this resulted in gifs ~1.9MB in size.

Note that an 8FPS version of the gif is 556KB in size and can be seen in
the initial commit for this PR.