Bugfix: Fix zooming with linked time by changing the definition of "clipped" (#5938)
Change the definition of "clipped" to "A portion of the time selection
cannot be shown"

## Motivation for features / changes
Currently when a the entirety of the time selection is outside of a
chart it is determined to be "clipped" and a section is chosen to be
shown.

When a card is "clipped" an icon is shown
![Screenshot from 2022-09-21
14-09-49](https://user-images.githubusercontent.com/78179109/191611033-34011978-c4d6-4fe4-88bf-417de62ada95.png)

The current behavior is somewhat confusing as it can lead to states
where the card does not show all the content but still does not register
as "clipped"
![Screenshot from 2022-09-21
14-15-28](https://user-images.githubusercontent.com/78179109/191611870-af2981fc-8097-437e-9b42-fdcda501b6fa.png)

The issue is then further compounded by "zooming"
![Screenshot from 2022-09-21
14-15-59](https://user-images.githubusercontent.com/78179109/191611977-685938b2-cb25-4d51-8fbf-b8c4e9c0df0a.png)

## Technical description of changes
I updated the `maybeClipTimeSelection` method to always return the
maximum time selection available within the bounds of the `minStep` and
`maxStep` provided and setting `clipped` to `true` if the
`timeSelection` has been changed.

## Screenshots of UI changes
Zooming with linked time enabled without range to include current step

![4c9091a2-ee78-4646-9266-548b67f9e6d1](https://user-images.githubusercontent.com/78179109/191847611-e676c47f-e1b8-477e-851d-df98b998ae5b.gif)

Zooming with linked time enabled to not include current step
![71900b0e-78ac-4f79-8a57-8d88d0b67134
(1)](https://user-images.githubusercontent.com/78179109/191847739-ae5d08b4-0257-4026-92d0-e71301803e09.gif)

Zooming to a subset of the time selection

![b2196686-1ab9-4ae9-9307-3865a89bdb74](https://user-images.githubusercontent.com/78179109/191846599-6732e602-b83f-4cdb-b188-42b3e9f59fd3.gif)

Zooming to a partial subset of the time selection

![8a8afbf5-b221-4bed-91ca-0dc5ffacf638](https://user-images.githubusercontent.com/78179109/191847133-3ae6e2e3-4e1c-4eb7-b034-eebf2da15a48.gif)

Zooming to a range including only a single whole number

![0f188df0-e82e-4c60-aae8-e1ad2716443c](https://user-images.githubusercontent.com/78179109/191847226-018e87af-c332-4f0b-b381-274a757e8c00.gif)

Zooming to a range including zero whole numbers

![f57f3f41-3f1b-4db5-82f5-01e86aba3928](https://user-images.githubusercontent.com/78179109/191847297-482e661d-1064-4303-a0d3-429568f74703.gif)

## Detailed steps to verify changes work correctly (as executed by you)
See gifs above

## Alternate designs / implementations considered
Not allowing zoom to effect the clipped status of a card.
I found the existing definition of clipped to be confusing and
unintuitive and zooming compounded that. Additionally it would have been
harder to maintain it working that way. For reference see [this
commit](https://github.com/rileyajones/tensorboard/commit/3b27f4e9803a3d1c5eceb0f65264118a9987d5da)
implementing it this way.