Range Selection: Simplify range selection in redux state (#5962)
* Motivation for features / changes
The redux state currently has two values used to represent whether range
selection is enabled (`stepSelectorRangeEnabled` and
`linkedTimeRangeEnabled`). Now that the UX around enabling range
selection is being changed to simply be a checkbox in the settings panel
this doesn't make a lot of sense.

This will enable #5956 will flush out the functionality for range
selection without linked time enabled.

* Technical description of changes
replace `stepSelectorRangeEnabled` and `linkedTimeRangeEnabled` with
`rangeSelectionEnabled` in redux state.

* Screenshots of UI changes
N/A

* Detailed steps to verify changes work correctly (as executed by you)
1) Start tensorboard
2) Go to
http://localhost:6006?allowRangeSelection&enableDataTable&enableLinkedTime
3) Enable "Link by Step" and add an end value
4) Assert that you now have two fobs (do not expect range selection to
be fully functional in without "Link By Step" enabled the rest of the
functionality is coming in #5956)

![image](https://user-images.githubusercontent.com/78179109/193884180-73205340-22d6-4745-8448-fdc080eaca8f.png)
5) Assert that "Enable Range Selection" is now enabled
6) Disable "Enable Range Selection"
7) Assert that the end fob disappears
8) Re-enable "Enable Range Selection"
9) Assert the end fob reappears.

* Alternate designs / implementations considered
Do not replace the two keys and just modify the reducers and selectors
to check both, this seems much messier.