Step Selector: Add action to sorting data table for metrics (#5964)
* Motivation for features / changes
We want to track some metrics to see if users are utilizing the sort by
feature in the data table. In order to do that we need to fire an action
when sorting. This PR fires that action. Note: currently nothing happens
on this action but, it sill get hooked into the metrics logging from the
internal code for our hosted services.

* Detailed steps to verify changes work correctly (as executed by you)
I created a reducer which fired from that action and console logged the
info just to ensure it fired. Then removed that reducer.