Step Selector: Update Redux Logic around enabling Range Selection (#5965)
# Motivation for features / changes
Currently "Enable Range Selection" DOES enable range selection, but it
does not create another fob so it doesn't appear to be enabled. This PR
handles the 1st part of the process, generating an end value in the
redux state. #5963 will handle the majority of the ui implementation.
We are planning to remove the right panel widget currently used to
enable range selection #5963.

# Technical description of changes
When range selection is enabled if and there is no current
`timeSelection` or `timeSelection` only has a single value generate a
new time selection value by finding the min / max step in all the
`timeSeriesData`

# Screenshots of UI changes
There is no ui effect to enabling range without linked time
## Disabling Step Selector Deselects "Enable Range Selection"

![a59eb425-11f4-4d08-8ef6-0e2c32e75188](https://user-images.githubusercontent.com/78179109/194379652-c2b8b2aa-fe97-4042-ae18-795ac96d2bde.gif)


## Enabling range with linked time enabled

![a1cc9fe4-2b70-4c38-9805-e62ffb121251](https://user-images.githubusercontent.com/78179109/194379411-1ab844aa-4c0b-45b3-95c0-8acf4b73c64b.gif)

## Detailed steps to verify changes work correctly (as executed by you)
**TLDR**: See gifs above
1) Start tensorboard
2) Go to
http://localhost:6006?enableLinkedTime=true&enableDataTable=true&allowRangeSelection=true#timeseries
3) Enable step selection and range selection
4) Assert only a single fob is shown on the scalar card and none on the
histogram card
5) Enable linked time
6) Assert two fobs are immediately shown on the scalar and histogram
cards
7) Disable range selection
8) Assert the second fob is removed from both cards.