Fix projector legend point coloring (#6021)
Fixes a regression where sprites are shown as black squares when first
loaded and when any sprite is deselected. This ultimately stemmed from a
null check (which colored the legend point white when null) being passed
over when `strictNullChecks` was introduced.

This allows `legendPointColorer` to take a null value. If `colorOption`
does not exist, it doesn't make sense to still go over the full dataSet
color assignment. Instead, return early to get `noSelectionColor`.

Googlers, see b/255947072.