Fix most tests that break with es2020 output. (#5991)
The migration to spec_bundle() will require we configure the TypeScript
compiler to output to JavaScript 'es2020'. There is one pattern of code
that we often use that will break:

in my_utilities.ts:
```
export function myFuncToSpyOn() {
  // Some code.
}
```

in my_test.ts:
```
import * as my_utilities from './my_utilities'

spyOn(my_utilities, 'myFuncToSpyOn').and.returnValue(1);
```

The Jasmine library will complain: 
`Error: <spyOn> : myFuncToSpyOn is not declared writable or has no
setter`

Jasmine is aware of the issue but can't provide a fix:
* https://github.com/jasmine/jasmine/issues/1942
* https://jasmine.github.io/pages/faq.html#module-spy

Instead they suggest we refactor our code so that the functions being
spied on are wrapped in some Class or other Object.

This PR does exactly that, for the most part doing the most simple of
wrapping:

now in my_utilities.ts:
```
// No longer exported.
function myFuncToSpyOn() {
  // Some code.
}

export const MyUtilities = {
  myFuncToSpyOn
}
```

now in my_test.ts:
```
import {MyUtilities} from './my_utilities'

spyOn(MyUtilities, 'myFuncToSpyOn').and.returnValue(1);
```

The change has a bit of bigger blast radius because not only do the
tests need to be changed but so does every other file (test or not) that
are using the utility functions.