Linked Time: Fix bug with zooming always maximizing end step (#6058)
## Motivation for features / changes
When changing the zoom on the scalar card with an the end fob the end
fob was always placed at the maximum possible step.
For Googlers b/259562961

## Technical description of changes
The `StepSelectorTimeSelection` derived by a subscription to a bunch of
other Observables. The main issue was that it simply never checked if an
existing end step existed. The code was also kinda hard to
read(:raised_hand_with_fingers_splayed: my fault) so I went ahead with a
bit of a cleanup there.

## Screenshots of UI changes
Before:
Zooming In:

![129f960a-2ee0-4762-98eb-875594b31b6f](https://user-images.githubusercontent.com/78179109/202584325-cdd164e7-2b04-4930-8de7-4c551ce85bcd.gif)

Resetting Zoom Level:

![00194071-2280-4d9b-953c-18740f7cad77](https://user-images.githubusercontent.com/78179109/202584465-4ddb3235-93b7-431b-92e1-16a7961d79e5.gif)


After:

![cf40658d-c57c-4d76-a267-9feb8b35f7fe](https://user-images.githubusercontent.com/78179109/202583150-6e7c662c-8230-4060-a289-7880dafda39a.gif)

## Detailed steps to verify changes work correctly (as executed by you)
1) Start tensorboard
2) Navigate to
http://localhost:6006?enableRangeSelection&enableDataTable
3) Place two fobs on the scalar card
4) Zoom in on a segment of the chart which contains both fobs (much
easier now that #5932 is closed)
5) Assert neither fob has moved
6) Reset the chart zoom level
7) Assert that neither fob has moved.